Move webpack config to subdirectory.
