+++ package.json
@@ -7,9 +7,9 @@
   "license": "MPL-2.0",
   "scripts": {
     "start": "web-ext run --source-dir build",
-    "build": "webpack --config webpack.dev.js",
-    "build:prod": "webpack --config webpack.prod.js",
-    "watch": "webpack --watch --config webpack.dev.js",
+    "build": "webpack --config webpack/dev.config.js",
+    "build:prod": "webpack --config webpack/prod.config.js",
+    "watch": "webpack --watch --config webpack/dev.config.js",
     "package": "web-ext build --source-dir build --overwrite-dest",
     "lint": "bin/run_lints.sh",
     "test": "bin/run_tests.py"
