+++ webpack/test.config.js
@@ -7,6 +7,7 @@
  */
 
 /* eslint-env node */
+/* eslint-disable import/no-extraneous-dependencies */
 
 const path = require('path');
 const webpack = require('webpack');
